mute deafen logic fixed, frame lost fixed

+++ app/src/main/java/com/github/jaykkumar01/watchparty_duo/utils/Base.java
@@ -31,7 +31,7 @@ public class Base {
     }
 
 
-    public static void sleep(int millis) {
+    public static void sleep(long millis) {
         try {
             Thread.sleep(millis);
         } catch (InterruptedException e) {
@@ -61,6 +61,7 @@ public class Base {
             return false;
         }
     }
+
 
 
 }
